added controllers in back and deleted get post page
