+++ front/src/App.js
@@ -5,7 +5,6 @@ import '../node_modules/bootstrap/dist/js/bootstrap.min.js'
 import {Routes, Route, Navigate} from 'react-router-dom'
 import Home from './pages/home/Home';
 import Post from './pages/post/CreatePost';
-import GetPost from './pages/post/GetPost';
 import SignUp from './pages/auth/SignUp';
 import Login from './pages/auth/Login';
 import ModifyPost from './pages/post/ModifyPost';
@@ -19,7 +18,6 @@ function App() {
         <Route path="/posts" element={<Home/>} />
         <Route path="/" element={<Login/>} />
         <Route path="/createpost" element={<Post/>} />
-        <Route path="/post/:id" element={<GetPost/>} />
         <Route path="/post/update/:id" element={<ModifyPost/>} />
         <Route path="/auth/login" element={<Login/>} />
         <Route path="/auth/signup" element={<SignUp/>} />
